+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this gear are documented in this file.
 
+## Unreleased
+
+* Updates to read in files with `utf-8-sig` to handle BOM encoding
+
 ## 1.1.0
 * Adds support for standalone forms submission
 * Strips leading zeros from PTID in error log name